+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ### Fixed
 
 - Set empty instead of null displayname and avatarUrl when deleting profile fields in old Matrix specs
+- Outbox did not retry on network error
 
 ### Security